worker/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
pika==1.3.0
psycopg2-binary==2.9.3
python-dotenv==0.20.0
sqlparse==0.4.2
